// webapp/ui/src/app/api-client/models/lottery-list-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ActiveLotterySummaryResponse } from '../models/active-lottery-summary-response';
import { HypeCountdownResponse } from '../models/hype-countdown-response';
import { LotteryEntryResponse } from '../models/lottery-entry-response';
export interface LotteryListResponse {
  active_lotteries?: Array<ActiveLotterySummaryResponse>;
  entries: Array<LotteryEntryResponse>;
  has_active_lottery?: boolean;
  hype_countdowns?: Array<HypeCountdownResponse>;
  latest_lotteries?: Array<ActiveLotterySummaryResponse>;
}